refactor(python): extract the Corpus pyclass into python/corpus.rs
Third step of the python.rs decomposition, and the first pyclass move — proves
the pattern the per-model legs will follow. The `Corpus` `#[pyclass]` + its
`Clone`/`#[pymethods]` impls move to `src/python/corpus.rs`.

Wiring details established here:
- The file is `corpus.rs` but the module is bound as `py_corpus` via `#[path]`,
  so it doesn't clash with the `use crate::corpus` import.
- `Corpus.inner` becomes `pub(crate)` (the ~31 model `fit` methods read it after
  `data.extract::<Corpus>()`); the other fields stay private behind getters.
- `build_corpus_from_docs` (still shared, used by every model) becomes
  `pub(super)` so corpus.rs can call it; `io_err` comes from `super::error`.

Pure move, no behavior change: public names identical; Corpus construction,
model-consuming-Corpus, and Corpus save/load all verified. clippy -D warnings +
fmt clean, cargo test --workspace (133 + 16), maturin --features python, pytest
(2782 passed, 26 skipped). faSTM untouched.

Next: one model leg (NMF) to confirm the model pattern, then the families.
